# solution/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AdsPlatform API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/advertisers/{advertiser_id}": {
      "get": {
        "tags": [
          "Advertisers"
        ],
        "summary": "Get Advertiser",
        "operationId": "get_advertiser_advertisers__advertiser_id__get",
        "parameters": [
          {
            "name": "advertiser_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Advertiser Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Advertiser"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/advertisers/bulk": {
      "post": {
        "tags": [
          "Advertisers"
        ],
        "summary": "Bulk Update",
        "operationId": "bulk_update_advertisers_bulk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/Advertiser"
                },
                "type": "array",
                "title": "Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Advertiser"
                  },
                  "type": "array",
                  "title": "Response Bulk Update Advertisers Bulk Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/advertisers/ml-scores": {
      "post": {
        "tags": [
          "Advertisers"
        ],
        "summary": "Set Score",
        "operationId": "set_score_advertisers_ml_scores_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MLScore"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/ads": {
      "get": {
        "tags": [
          "Ads"
        ],
        "summary": "Get Ad",
        "operationId": "get_ad_ads_get",
        "parameters": [
          {
            "name": "client_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Client Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ad"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ads/{ad_id}/click": {
      "post": {
        "tags": [
          "Ads"
        ],
        "summary": "Click Ad",
        "operationId": "click_ad_ads__ad_id__click_post",
        "parameters": [
          {
            "name": "ad_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ad Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_click_ad_ads__ad_id__click_post"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Ad": {
        "properties": {
          "ad_id": {
            "type": "string",
            "format": "uuid",
            "title": "Ad Id"
          },
          "ad_title": {
            "type": "string",
            "title": "Ad Title"
          },
          "ad_text": {
            "type": "string",
            "title": "Ad Text"
          },
          "advertiser_id": {
            "type": "string",
            "format": "uuid",
            "title": "Advertiser Id"
          }
        },
        "type": "object",
        "required": [
          "ad_id",
          "ad_title",
          "ad_text",
          "advertiser_id"
        ],
        "title": "Ad"
      },
...
      "Advertiser": {
        "properties": {
          "advertiser_id": {
            "type": "string",
            "format": "uuid",
            "title": "Advertiser Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "advertiser_id",
          "name"
        ],
        "title": "Advertiser"
      },
      "Body_click_ad_ads__ad_id__click_post": {
        "properties": {
          "client_id": {
            "type": "string",
            "format": "uuid",
            "title": "Client Id"
          }
        },
        "type": "object",
        "required": [
          "client_id"
        ],
        "title": "Body_click_ad_ads__ad_id__click_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MLScore": {
        "properties": {
          "client_id": {
            "type": "string",
            "format": "uuid",
            "title": "Client Id"
          },
          "advertiser_id": {
            "type": "string",
            "format": "uuid",
            "title": "Advertiser Id"
          },
          "score": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "client_id",
          "advertiser_id",
          "score"
        ],
        "title": "MLScore"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}